+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v4.0.0 2020-10-13
 
 - `bind` method rejects on error. Fixes #42.
+- `recv` method resolves on `message` event. Fixes #39.
 - Requires Node >= 10.
 - Uses eslint instead of tslint.